requirements.txt
python-dotenv
patchright
playwright>=1.49.0
aiohttp>=3.11.0
beautifulsoup4>=4.12.0
requests>=2.31.0